move versions component to dev folder
